+++ test/unit/messagesTest.ts
@@ -366,4 +366,50 @@ describe('Messages', () => {
       expect(error.actions[0]).to.equal('from prefix');
     });
   });
+  describe('createWarning', () => {
+    it('creates warning with actions', () => {
+      const messages = new Messages('myBundle', Messages.getLocale(), msgMap);
+      const warning = messages.createWarning('msg1');
+
+      expect(warning.name).to.equal('Msg1Warning');
+      expect(warning.message).to.equal(testMessages.msg1);
+      expect(warning.actions.length).to.equal(1);
+      expect(warning.actions[0]).to.equal(testMessages.msg2);
+    });
+
+    it('creates warning with removed warning prefix', () => {
+      msgMap.set('warning.msg1', msgMap.get('msg1'));
+      msgMap.set('warning.msg1.actions', ['from prefix']);
+      const messages = new Messages('myBundle', Messages.getLocale(), msgMap);
+      const warning = messages.createWarning('warning.msg1');
+
+      expect(warning.name).to.equal('Msg1Warning');
+      expect(warning.message).to.equal(testMessages.msg1);
+      expect(warning.actions.length).to.equal(1);
+      expect(warning.actions[0]).to.equal('from prefix');
+    });
+  });
+  describe('createInfo', () => {
+    it('creates info with actions', () => {
+      const messages = new Messages('myBundle', Messages.getLocale(), msgMap);
+      const info = messages.createInfo('msg1');
+
+      expect(info.name).to.equal('Msg1Info');
+      expect(info.message).to.equal(testMessages.msg1);
+      expect(info.actions.length).to.equal(1);
+      expect(info.actions[0]).to.equal(testMessages.msg2);
+    });
+
+    it('creates info with removed info prefix', () => {
+      msgMap.set('info.msg1', msgMap.get('msg1'));
+      msgMap.set('info.msg1.actions', ['from prefix']);
+      const messages = new Messages('myBundle', Messages.getLocale(), msgMap);
+      const info = messages.createInfo('info.msg1');
+
+      expect(info.name).to.equal('Msg1Info');
+      expect(info.message).to.equal(testMessages.msg1);
+      expect(info.actions.length).to.equal(1);
+      expect(info.actions[0]).to.equal('from prefix');
+    });
+  });
 });